pyproject.toml
[project]
name = "live-stock-app"
version = "0.1.0"
description = "Stock analysis & recommendation app powered by Zerodha Kite Connect"
requires-python = ">=3.11"

dependencies = [
    "kiteconnect>=5.0.0",
    "pandas>=2.0",
    "ta>=0.11.0",
    "pydantic>=2.0",
    "mcp>=1.0",
    "python-dotenv>=1.0",
    "rich>=13.0",
    "click>=8.0",
    "httpx>=0.27",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "yfinance>=0.2.40",
    "python-multipart>=0.0.9",
    "truststore>=0.10",
    "statsmodels>=0.14",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "apscheduler>=3.10",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
stock-app = "src.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
